Fix missing ModuleService at initialization
