+++ frontend/src/app/app.module.ts
@@ -1,6 +1,6 @@
 // Angular core
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, APP_INITIALIZER } from '@angular/core';
+import { NgModule, APP_INITIALIZER, Injector } from '@angular/core';
 
 import {
   HttpClientModule,
@@ -38,11 +38,11 @@ import { IntroductionComponent } from '../custom/components/introduction/introdu
 import { AuthService } from './components/auth/auth.service';
 import { CookieService } from 'ng2-cookies';
 import { ChartsModule } from 'ng2-charts';
-import {
-  AuthGuard,
-  ModuleGuardService,
-  PublicAccessGuard,
-} from '@geonature/routing/routes-guards.service';
+
+// PublicAccessGuard,
+
+import { AuthGuard } from '@geonature/routing/auth-guard.service';
+import { ModuleGuardService } from '@geonature/routing/module-guard.service';
 import { ModuleService } from './services/module.service';
 import { CruvedStoreService } from './GN2CommonModule/service/cruved-store.service';
 import { SideNavService } from './components/sidenav-items/sidenav-service';
@@ -50,7 +50,8 @@ import { SideNavService } from './components/sidenav-items/sidenav-service';
 import { MyCustomInterceptor } from './services/http.interceptor';
 import { UnauthorizedInterceptor } from './services/unauthorized.interceptor';
 import { GlobalSubService } from './services/global-sub.service';
-
+import { tap } from 'rxjs/operators';
+import { RoutingService } from './routing/routing.service';
 export function createTranslateLoader(http: HttpClient) {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
 }
@@ -60,9 +61,19 @@ import { UserDataService } from './userModule/services/user-data.service';
 import { APP_CONFIG_TOKEN, AppConfig } from '@geonature_config/app.config';
 import { Router } from '@angular/router';
 
-export function getModulesAndInitRouting(moduleService: ModuleService, router: Router) {
+export function getModulesAndInitRouting(injector) {
   return () => {
-    return moduleService.fetchModulesAndSetRouting().toPromise();
+    // return moduleService.fetchModulesAndSetRouting().toPromise();
+    const moduleService = injector.get(ModuleService);
+    const routingService = injector.get(RoutingService);
+    return moduleService
+      .loadModules()
+      .pipe(
+        tap((modules) => {
+          routingService.loadRoutes(modules);
+        })
+      )
+      .toPromise();
   };
 }
 
@@ -111,7 +122,6 @@ export function getModulesAndInitRouting(moduleService: ModuleService, router: R
     SideNavService,
     CruvedStoreService,
     UserDataService,
-    PublicAccessGuard,
     { provide: APP_CONFIG_TOKEN, useValue: AppConfig },
     { provide: HTTP_INTERCEPTORS, useClass: MyCustomInterceptor, multi: true },
     { provide: HTTP_INTERCEPTORS, useClass: UnauthorizedInterceptor, multi: true },
@@ -119,7 +129,7 @@ export function getModulesAndInitRouting(moduleService: ModuleService, router: R
     {
       provide: APP_INITIALIZER,
       useFactory: getModulesAndInitRouting,
-      deps: [ModuleService, Router],
+      deps: [Injector],
       multi: true,
     },
   ],
