+++ frontend/src/app/routing/app-routing.module.ts
@@ -1,13 +1,9 @@
-// WARNING: this file was auto generated by GeoNature command line.
-// See Jinja template with ".sample" file extension in same directory.
-
-import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { HomeContentComponent } from '../components/home-content/home-content.component';
 import { PageNotFoundComponent } from '../components/page-not-found/page-not-found.component';
-import { AuthGuard, ModuleGuardService } from '@geonature/routing/routes-guards.service';
-
+import { AuthGuard } from '@geonature/routing/auth-guard.service';
+import { ModuleGuardService } from '@geonature/routing/module-guard.service';
 import { SignUpGuard } from '@geonature/modules/login/routes-guard.service';
 import { SignUpComponent } from '../modules/login/sign-up/sign-up.component';
 
@@ -16,7 +12,6 @@ import { NewPasswordComponent } from '../modules/login/new-password/new-password
 
 import { LoginComponent } from '../modules/login/login/login.component';
 import { NavHomeComponent } from '../components/nav-home/nav-home.component';
-import { MediasTestComponent } from '@geonature_common/form/media/medias-test.component';
 
 const defaultRoutes: Routes = [
   {
