+++ frontend/src/app/services/module.service.ts
@@ -3,7 +3,6 @@ import { Router } from '@angular/router';
 import { DataFormService } from '@geonature_common/form/data-form.service';
 import { Observable, of, BehaviorSubject } from 'rxjs';
 import { catchError, tap } from 'rxjs/operators';
-import { ModuleGuardService } from '../routing/routes-guards.service';
 
 @Injectable()
 export class ModuleService {
@@ -24,34 +23,10 @@ export class ModuleService {
 
   constructor(private _api: DataFormService, private _router: Router) {}
 
-  fetchModulesAndSetRouting(): Observable<any[]> {
-    // see CruvedStoreService.fetchCruved comments about the catchError
+  loadModules(): Observable<any[]> {
     return this._api.getModulesList([]).pipe(
       catchError((err) => of([])), // TODO: error MUST be handled in case we are logged! (typically, api down)
       tap((modules) => {
-        const routingConfig = this._router.config;
-        modules.forEach((module) => {
-          if (module.ng_module) {
-            const moduleConfig = {
-              path: module.module_path,
-              loadChildren: () =>
-                import(
-                  /* webpackInclude: /\/external_modules\/[^/]*\/frontend\// */
-                  `../../../../external_modules/${module.ng_module}/frontend/app/gnModule.module`
-                ).then((m) => m.GeonatureModule),
-              canActivate: [ModuleGuardService],
-              data: {
-                module_code: module.module_code,
-              },
-            };
-            // insert at the begining otherwise pagenotfound component is first matched
-            const basePathIndex = routingConfig.findIndex((route) => {
-              return route.path === '';
-            });
-            routingConfig[basePathIndex].children.unshift(moduleConfig);
-          }
-        });
-        this._router.resetConfig(routingConfig);
         this.modules = modules;
       })
     );
